refactor: isolate child launch context
